Add null support to aggregates

+++ presto-main/src/test/java/com/facebook/presto/operator/aggregation/DoubleSequenceCursor.java
@@ -85,7 +85,7 @@ public class DoubleSequenceCursor
     {
         checkReadablePosition();
         return getTupleInfo().builder()
-                .append(current)
+                .append((double) current)
                 .build();
     }
